+++ js/react/es5/ImageGallery.js
@@ -45,11 +45,21 @@ try {
   window.removeEventListener('testPassive', null, opts);
 } catch (e) {}
 
+var isMobile = function () {
+  var ua = window.navigator.userAgent.toLowerCase();
+  return ua.indexOf('android') > -1 || ua.indexOf('iphone') > -1 || ua.indexOf('ipad') > -1;
+}();
+
+var supportsAvif = null;
+var supportsWebp = null;
+
 function BasicImage(props) {
   return React.createElement('img', {
     src: props.thumbnail,
     alt: props.title,
-    onClick: props.onClick
+    tabIndex: props.tabIndex,
+    onClick: props.onClick,
+    onKeyDown: props.onKeyDown
   });
 }
 
@@ -65,23 +75,30 @@ var ImageGallery = function (_React$Component) {
 
     _this = _super.call(this, props);
     _this.onClick = _this.onClick.bind(_assertThisInitialized(_this));
+    _this.onKeyDown = _this.onKeyDown.bind(_assertThisInitialized(_this));
     _this.handleDocKeyDown = _this.handleDocKeyDown.bind(_assertThisInitialized(_this));
     _this.preloadNextImages = _this.preloadNextImages.bind(_assertThisInitialized(_this));
+    _this.showOverlay = _this.showOverlay.bind(_assertThisInitialized(_this));
     _this.hideOverlay = _this.hideOverlay.bind(_assertThisInitialized(_this));
     _this.changeImage = _this.changeImage.bind(_assertThisInitialized(_this));
+    _this.svgForwardButton = '<svg width="13" height="22" xmlns="http://www.w3.org/2000/svg"><path d="M3.4.6L12 9c.4.4.6 1 .6 1.5a2 2 0 01-.6 1.5l-8.5 8.5a2 2 0 01-2.8-2.8l7.2-7.2L.6 3.4A2 2 0 013.4.6z" fill="#fff" fill-rule="evenodd"/></svg>';
+    _this.svgBackButton = '<svg width="13" height="22" xmlns="http://www.w3.org/2000/svg"><path d="M9 .6L.7 9a2 2 0 00-.6 1.5c0 .5.2 1.1.6 1.5L9 20.6a2 2 0 002.8-2.8l-7.2-7.2L12 3.4A2 2 0 009.1.6z" fill="#fff" fill-rule="evenodd"/></svg>';
     _this.overlayStyleId = 'image-gallery-css';
-    _this.overlayStyleCss = "\n            body.blur { filter: blur(3px); }\n\n            .image-gallery-overlay {\n                position: fixed;\n                top: 0;\n                left: 0;\n                right: 0;\n                bottom: 0;\n                background-color: rgba(255,255,255,.8);\n                cursor: pointer;\n                display: flex;\n                justify-content: center;\n                align-items: center;\n                flex-direction: column;\n            }\n\n            .image-gallery-overlay .image-gallery-loading {\n                font-weight: bold;\n                padding: 10px 20px;\n                background-color: rgba(255,255,255,.4);\n                position: absolute;\n            }\n\n            .image-gallery-overlay img {\n                max-width: 100%;\n                max-height: 100%;\n                flex-shrink: 0;\n            }\n\n            .image-gallery-overlay div {\n                position: absolute;\n                bottom: 0;\n                left: 0;\n                right: 0;\n                z-index: 2;\n                font-weight: bold;\n                display: flex;\n                justify-content: space-between;\n                width: 100%;\n            }\n\n            .image-gallery-overlay div span {\n                padding: 10px 20px;\n                background-color: rgba(255,255,255,.4);\n            }\n\n            @media (min-width: 1300px) {\n                .image-gallery-overlay div {\n                    left: calc((100% - 1300px) /2);\n                    right: auto;\n                    max-width: 1300px;\n                }\n            }\n        ";
+    _this.overlayStyleCss = "\n            body.blur { filter: blur(3px); }\n\n            .image-gallery-overlay {\n                position: fixed;\n                top: 0;\n                left: 0;\n                right: 0;\n                bottom: 0;\n                background-color: rgba(255,255,255,.8);\n                cursor: pointer;\n                display: flex;\n                justify-content: center;\n                align-items: center;\n                flex-direction: column;\n            }\n\n            .image-gallery-overlay .image-gallery-loading {\n                font-weight: bold;\n                padding: 10px 20px;\n                background-color: rgba(255,255,255,.4);\n                position: absolute;\n            }\n\n            .image-gallery-overlay img {\n                max-width: 100%;\n                max-height: 100%;\n                flex-shrink: 0;\n            }\n\n            .image-gallery-overlay div {\n                position: absolute;\n                bottom: 0;\n                left: 0;\n                right: 0;\n                z-index: 2;\n                font-weight: bold;\n                display: flex;\n                justify-content: space-between;\n                width: 100%;\n            }\n\n            .image-gallery-overlay div span {\n                padding: 10px 20px;\n                background-color: rgba(255,255,255,.4);\n            }\n\n            .image-gallery-overlay .btn-previous,\n            .image-gallery-overlay .btn-next {\n                display: block;\n                position: absolute;\n                height: 40px;\n                width: 40px;\n                opacity: .7;\n                background-repeat: no-repeat;\n                background-position: center;\n                padding: 0;\n                margin: 15px;\n                background-color: rgba(0,0,0,.5);\n                border-radius: 50%;\n            }\n            .image-gallery-overlay .btn-previous { left: 0; background-position-x: 12px; background-image: url(\"data:image/svg+xml;base64,".concat(btoa(_this.svgBackButton), "\"); }\n            .image-gallery-overlay .btn-next { right: 0; background-position-x: 15px; background-image: url(\"data:image/svg+xml;base64,").concat(btoa(_this.svgForwardButton), "\"); }\n\n            .image-gallery-overlay.mobile .btn-previous,\n            .image-gallery-overlay.mobile .btn-next,\n            .image-gallery-overlay.keyboard .btn-previous,\n            .image-gallery-overlay.keyboard .btn-next {\n                display: none;\n            }\n\n            @media (min-width: 1300px) {\n                .image-gallery-overlay div {\n                    left: calc((100% - 1300px) /2);\n                    right: auto;\n                    max-width: 1300px;\n                }\n            }\n\n            @media screen and (-ms-high-contrast: active), screen and (-ms-high-contrast: none) {\n                .image-gallery-overlay .image-gallery-loading,\n                .image-gallery-overlay .btn-previous,\n                .image-gallery-overlay .btn-next { margin-top: calc((100vh /2) - 35px); }\n            }\n        ");
     _this.imageIndex = null;
     _this.overlay = null;
     _this.overlayImg = null;
     _this.overlayTitle = null;
     _this.overlayIndex = null;
     _this.overlayLoading = null;
+    _this.overlayBackButton = null;
+    _this.overlayFowardButton = null;
     _this.touchStartX = null;
     _this.loadingTimeoutId = null;
     _this.loadingText = props.loadingText ? props.loadingText : 'Loading...';
     _this.loadingTimeout = props.loadingTimeout ? props.loadingTimeout : 2000;
     _this.loadedImages = new Set();
+    _this.startedFromKeyboard = false;
     _this.state = {
       images: props.images
     };
@@ -106,7 +123,19 @@ var ImageGallery = function (_React$Component) {
         }
       }
 
-      this.showOverlay();
+      this.checkSupportedFormats();
+    }
+  }, {
+    key: "onKeyDown",
+    value: function onKeyDown(e) {
+      if (e.key === ' ' || e.key === 'Spacebar') {
+        if (this.overlay === null) {
+          this.startedFromKeyboard = true;
+          this.onClick(e);
+        }
+
+        e.preventDefault();
+      }
     }
   }, {
     key: "loadCss",
@@ -121,20 +150,82 @@ var ImageGallery = function (_React$Component) {
       }
     }
   }, {
+    key: "checkSupportedFormats",
+    value: function checkSupportedFormats() {
+      if (supportsAvif !== null && supportsWebp !== null) {
+        this.showOverlay();
+        return;
+      }
+
+      var showOverlay = this.showOverlay;
+
+      function checkStatus() {
+        if (supportsAvif !== null && supportsWebp !== null) {
+          showOverlay();
+          return;
+        }
+      }
+
+      var imgAvif = new Image();
+
+      imgAvif.onload = function () {
+        supportsAvif = imgAvif.width === 1 && imgAvif.height === 1;
+        checkStatus();
+      };
+
+      imgAvif.onerror = function () {
+        supportsAvif = false;
+        checkStatus();
+      };
+
+      imgAvif.src = 'data:image/avif;base64,AAAAIGZ0eXBhdmlmAAAAAGF2aWZtaWYxbWlhZk1BMUIAAADybWV0YQAAAAAAAAAoaGRscgAAAAAAAAAAcGljdAAAAAAAAAAAAAAAAGxpYmF2aWYAAAAADnBpdG0AAAAAAAEAAAAeaWxvYwAAAABEAAABAAEAAAABAAABGgAAABoAAAAoaWluZgAAAAAAAQAAABppbmZlAgAAAAABAABhdjAxQ29sb3IAAAAAamlwcnAAAABLaXBjbwAAABRpc3BlAAAAAAAAAAEAAAABAAAAEHBpeGkAAAAAAwgICAAAAAxhdjFDgQ0MAAAAABNjb2xybmNseAACAAIAAYAAAAAXaXBtYQAAAAAAAAABAAEEAQKDBAAAACJtZGF0EgAKCBgABogQEAwgMgwYAAAAUAAAALASmpg=';
+      var imgWebP = new Image();
+
+      imgWebP.onload = function () {
+        supportsWebp = imgWebP.width === 1 && imgWebP.height === 1;
+        checkStatus();
+      };
+
+      imgWebP.onerror = function () {
+        supportsWebp = false;
+        checkStatus();
+      };
+
+      imgWebP.src = 'data:image/webp;base64,UklGRiIAAABXRUJQVlA4IBYAAAAwAQCdASoBAAEADsD+JaQAA3AAAAAA';
+    }
+  }, {
     key: "showOverlay",
     value: function showOverlay() {
       var _this2 = this;
 
-      var imageSrc = this.state.images[this.imageIndex].image;
+      var imageSrc = this.getImage(this.state.images[this.imageIndex]);
       var imageTitle = this.state.images[this.imageIndex].title;
       this.loadCss();
       this.overlay = document.createElement('div');
-      this.overlay.className = 'image-gallery-overlay';
+      this.overlay.className = 'image-gallery-overlay' + (isMobile ? ' mobile' : '') + (this.startedFromKeyboard ? ' keyboard' : '');
       this.overlayLoading = document.createElement('span');
       this.overlayLoading.className = 'image-gallery-loading';
       this.overlayLoading.textContent = this.loadingText;
       this.overlayLoading.setAttribute('hidden', '');
       this.overlay.appendChild(this.overlayLoading);
+      this.overlayBackButton = document.createElement('span');
+      this.overlayBackButton.className = 'btn-previous';
+      this.overlayBackButton.setAttribute('role', 'button');
+
+      this.overlayBackButton.onclick = function () {
+        _this2.changeImage('left');
+      };
+
+      this.overlayFowardButton = document.createElement('span');
+      this.overlayFowardButton.className = 'btn-next';
+      this.overlayFowardButton.setAttribute('role', 'button');
+
+      this.overlayFowardButton.onclick = function () {
+        _this2.changeImage('right');
+      };
+
+      this.overlay.appendChild(this.overlayBackButton);
+      this.overlay.appendChild(this.overlayFowardButton);
       this.overlayImg = document.createElement('img');
       this.overlayImg.addEventListener('load', function () {
         _this2.loadedImages.add(imageSrc);
@@ -162,6 +253,19 @@ var ImageGallery = function (_React$Component) {
       document.documentElement.appendChild(this.overlay);
       document.querySelector('body').classList.add('blur');
       this.startLoadingTimer();
+    }
+  }, {
+    key: "getImage",
+    value: function getImage(image) {
+      if (image.image_avif !== undefined && supportsAvif) {
+        return image.image_avif;
+      }
+
+      if (image.image_webp !== undefined && supportsWebp) {
+        return image.image_webp;
+      }
+
+      return image.image;
     }
   }, {
     key: "startLoadingTimer",
@@ -201,6 +305,10 @@ var ImageGallery = function (_React$Component) {
           if (e.clientY >= screen25pct && e.clientY <= screenBottom) {
             return;
           }
+        }
+
+        if (e.target === _this4.overlayBackButton || e.target === _this4.overlayFowardButton) {
+          return;
         }
 
         _this4.hideOverlay();
@@ -252,11 +360,14 @@ var ImageGallery = function (_React$Component) {
     value: function hideOverlay() {
       this.clearLoadingTimer();
       this.overlay.parentNode.removeChild(this.overlay);
+      this.overlayBackButton = null;
+      this.overlayFowardButton = null;
       this.overlayLoading = null;
       this.overlayIndex = null;
       this.overlayTitle = null;
       this.overlayImg = null;
       this.overlay = null;
+      this.startedFromKeyboard = false;
       document.removeEventListener('keydown', this.handleDocKeyDown);
       document.querySelector('body').classList.remove('blur');
     }
@@ -278,7 +389,7 @@ var ImageGallery = function (_React$Component) {
       }
 
       if (indexLeft !== this.imageIndex) {
-        var srcLeft = this.state.images[indexLeft].image;
+        var srcLeft = this.getImage(this.state.images[indexLeft]);
 
         if (srcLeft && !this.loadedImages.has(srcLeft)) {
           var imgLeft = new Image();
@@ -298,7 +409,7 @@ var ImageGallery = function (_React$Component) {
       }
 
       if (indexRight !== this.imageIndex) {
-        var srcRight = this.state.images[indexRight].image;
+        var srcRight = this.getImage(this.state.images[indexRight]);
 
         if (srcRight && !this.loadedImages.has(srcRight)) {
           var imgRight = new Image();
@@ -324,7 +435,7 @@ var ImageGallery = function (_React$Component) {
 
       var imageTitle = this.state.images[this.imageIndex].title;
       this.overlayImg.src = '';
-      this.overlayImg.src = this.state.images[this.imageIndex].image;
+      this.overlayImg.src = this.getImage(this.state.images[this.imageIndex]);
       this.overlayTitle.textContent = imageTitle;
       this.overlayTitle.style.display = imageTitle ? '' : 'none';
       this.overlayIndex.textContent = "".concat(this.imageIndex + 1, "/").concat(imageCount);
@@ -345,10 +456,20 @@ var ImageGallery = function (_React$Component) {
         }
       }
 
+      var tabIndex = parseInt(this.props.tabIndex, 10);
+      var useTabIndex = window.isFinite(tabIndex);
       return this.state.images.map(function (image) {
-        return React.cloneElement(template, Object.assign({}, image, {
-          onClick: _this6.onClick
-        }));
+        var imageAttr = {
+          onClick: _this6.onClick,
+          onKeyDown: _this6.onKeyDown
+        };
+
+        if (useTabIndex) {
+          imageAttr.tabIndex = tabIndex;
+          tabIndex++;
+        }
+
+        return React.cloneElement(template, Object.assign({}, image, imageAttr));
       });
     }
   }]);